Improved engine particles, imo.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@785 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Particles/EngineParticle.h
@@ -6,7 +6,7 @@
 class EngineParticle : public Particle {
    public:
       EngineParticle(Point3D* p, Vector3D* v, float life, float r, float g, float b, const GameState* _gameState);
-      static void Add(Point3D* pos, Vector3D* vec, int color, const GameState*);
+      static void Add(Point3D* pos, Vector3D* vec, double color, const GameState*);
       virtual void update(double timeDifference);
 
 };
